+++ src/activityManager.js
@@ -77,9 +77,13 @@ _.extend(ActivityManager.prototype, {
     },
 
     reset: function() {
-        this._displayRegion && this._displayRegion.close();
-        this._currentActivity = null;
-        this._currentPlace = null;
+        if (this._displayRegion) {                                                       
+            this._displayRegion.close();                                         
+            this._displayRegion.hide();                                          
+        }                                                                        
+        this._deactivateCurrentActivity();                                       
+        this._currentActivity = null;                                            
+        this._currentPlace = null; 
     },
 
     // Loads an activity from a place, returns a promise that indicates when
@@ -95,8 +99,11 @@ _.extend(ActivityManager.prototype, {
         // No match was found so we reset the activity manager.
         if (!match) {
             this.reset();
+            this._displayRegion.hide();
             return resolvedPromise;
         }
+
+        this._displayRegion.show();
         
         // If the new place equals the last one, no activity is loaded.
         if (place.equals(this._currentPlace)) {
@@ -110,7 +117,7 @@ _.extend(ActivityManager.prototype, {
         activity = this._createActivity(match.Activity, place);
 
         // Perform some deactivation tasks over the previous activity.
-        this._deactivate(this._currentActivity);
+        this._deactivateCurrentActivity();
 
         // Store the new activity as the current one.
         this._currentActivity = activity;
@@ -149,6 +156,10 @@ _.extend(ActivityManager.prototype, {
         return this._promise;
     },
 
+    _deactivateCurrentActivity: function() {
+        this._deactivate(this._currentActivity);
+    },
+
     _deactivate: function(activity) {
         var _self = this;
 
@@ -177,7 +188,7 @@ _.extend(ActivityManager.prototype, {
 
     showView: function(view) {
         if (this._displayRegion) {
-            this._displayRegion.show(view);
+            this._displayRegion.display(view);
         }
     },
 
@@ -199,6 +210,7 @@ var ProtectedDisplay = function(activityManager) {
     // storing the current activity to compare later
     this.activity = activityManager.getCurrentActivity();
     this.activityManager = activityManager;
+    this.region = this.activityManager.getDisplayRegion();
 
     this._deferred = $.Deferred();
     // Promise to inform the activity manager that the activity has finished
@@ -207,12 +219,19 @@ var ProtectedDisplay = function(activityManager) {
 }
 
 _.extend(ProtectedDisplay.prototype, { 
-    setView: function(view) {
+    /**
+     * @param view {Backbone.View|String|Element}
+     * @param resolve {Boolean} optional, determines if the deferred should be resolved.
+     */
+    setView: function(view, resolve) {
         var activityManager = this.activityManager;
 
         if (this.activity == activityManager.getCurrentActivity()) {
             activityManager.showView(view);
-            this._deferred.resolve(activityManager._currentPlace);
+
+            if (typeof resolve === "undefined" || resolve === true) {
+                this._deferred.resolve(activityManager._currentPlace);
+            }
         }
     },